packages/telemetry/python/pyproject.toml
[project]
name = "latitude-telemetry"
version = "3.0.0a0"
description = "Latitude Telemetry for Python"
authors = [{ name = "Latitude Data SL", email = "hello@latitude.so" }]
maintainers = [{ name = "Latitude Data SL", email = "hello@latitude.so" }]
readme = "README.md"
license = "MIT"
urls.repository = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/python"
urls.homepage = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/python#readme"
urls.documentation = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/python#readme"
requires-python = ">=3.11, <3.15"
dependencies = [
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
    "opentelemetry-semantic-conventions-ai>=0.4.2",
    "openinference-semantic-conventions>=0.1.12",
    "opentelemetry-instrumentation-alephalpha>=0.50.1",
    "opentelemetry-instrumentation-anthropic>=0.50.1",
    "opentelemetry-instrumentation-bedrock>=0.50.1",
    "opentelemetry-instrumentation-cohere>=0.50.1",
    "opentelemetry-instrumentation-crewai>=0.50.1",
    "openinference-instrumentation-dspy>=0.1.31",
    "opentelemetry-instrumentation-google-generativeai>=0.50.1",
    "opentelemetry-instrumentation-groq>=0.50.1",
    "opentelemetry-instrumentation-haystack>=0.50.1",
    "opentelemetry-instrumentation-langchain>=0.50.1",
    "openinference-instrumentation-litellm>=0.1.6",
    "opentelemetry-instrumentation-llamaindex>=0.50.1",
    "opentelemetry-instrumentation-mistralai>=0.50.1",
    "opentelemetry-instrumentation-ollama>=0.50.1",
    "opentelemetry-instrumentation-openai>=0.50.1",
    "opentelemetry-instrumentation-replicate>=0.50.1",
    "opentelemetry-instrumentation-sagemaker>=0.50.1",
    "opentelemetry-instrumentation-together>=0.50.1",
    "opentelemetry-instrumentation-threading>=0.50b0",
    "opentelemetry-instrumentation-transformers>=0.50.1",
    "opentelemetry-instrumentation-vertexai>=0.50.1",
    "opentelemetry-instrumentation-watsonx>=0.50.1",
    "pydantic>=2.12",
    "typing-extensions>=4.12.2",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest-xdist>=3.8.0",
    "pytest>=8.4.2",
    "pyright>=1.1.405",
    "ruff>=0.13.0",
    "sh>=2.2.2",
    "openai>=1.59.7",
    # Note: 0.41.0 onwards made Prompt Caching generally available
    # which is not compatible with the instrumentation
    "anthropic==0.40.0",
    "litellm[proxy]>=1.58.2",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnnecessaryIsInstance = false
reportPrivateUsage = false

[tool.ruff]
target-version = "py311"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["B", "C4", "E", "F", "I", "W", "UP"]
ignore = [
    "F401",
    "F403",
    # Needed because unnecessary str() on field aliases are needed
    # https://docs.pydantic.dev/2.8/concepts/fields/#field-aliases
    "UP018",
    # Needed because typing.List and typing.Dict are semi-deprecated
    # in new Python versions but we want to maintain compatibility
    "UP006",
    "UP035",
    "E402",  # We allow import not at top of file to not import every framework package
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
addopts = "-p no:warnings -n auto"
xfail_strict = true
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
